vcnc-web: Factored a 'NumberWidget' out of 'StorageEfficiencyNumber'. Made the font size responsive to viewport width (but alas not the div width).
